% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metastate.R
\name{add_metastate_model}
\alias{add_metastate_model}
\title{Create a state transition framework}
\usage{
add_metastate_model(daps, state, nodes, transitions)
}
\arguments{
\item{daps}{A \code{\link{daps}}-class object to add a state-transition framework to.}

\item{state}{A character vector of state names. Must be the same nonzero
lenth as \code{nodes} and \code{transitions}.}

\item{nodes}{A list the same nonzero length as \code{state} and \code{transitions}. It
must contain character vectors, each naming the nodes (or variables) that
should be predicted under the corresponding \code{state}.}

\item{transitions}{A list the same nonzero length as \code{state} and \code{nodes}. It
must contain one-sided formulas, each containing an expression that
produces one of the elements of \code{state} when evaluated in the context of a
row of temporal data (provided to \code{\link[=simulate]{simulate()}}).}
}
\value{
The input \code{daps} with the added component \code{$metastate_model}, an
object of class \code{daps_metastate_model}.
}
\description{
Adds a state transition framework to a \code{daps} object.
}
