% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_specification.R
\name{add_models}
\alias{add_models}
\title{Add individual variable models to a \code{daps} object.}
\usage{
add_models(daps, ..., daps_model_table = NULL)
}
\arguments{
\item{daps}{The \code{\link{daps}}-class object to add individual variable models to.}

\item{...}{Individual variable models. This can by any combination of
two-sided \code{\link[base:formula]{formula}}\code{s} and calls to modeling functions.

The left side of each formula must specify a single variable that will
appear in the data. The right side will be extracted and saved to be
evaluated in the context of a row of the data during simulation (i.e., in
\code{\link[=simulate]{simulate()}}) as needed.

Calls to modeling functions are not evaluated until model training (i.e.,
in \code{\link[=train]{train()}}). They must be written without a \code{data} argument, as this
argument will be filled in during training. The left side of the \code{formula}
argument of modeling function calls must specify a single variable that
will appear in the data.

Multiple formulas/model calls can specify the same variable. These should
be listed in descending order of preference (which may usually also mean
descending order of complexity).

These arguments are not evaluated but quoted.

Ignored if \code{daps_model_table} is not \code{NULL}.}

\item{daps_model_table}{An object of class \code{daps_model_table} previously
created. If not \code{NULL}, anything specified in \code{...} will be ignored.}
}
\value{
The input \code{daps} with the added component \verb{$model_table}, an object
of class \code{daps_model_table}.
}
\description{
It is expected that multiple entries will apply to the same variable. They
should be entered in descending order of preference.
}
