% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DAPS-package.R
\docType{package}
\name{DAPS-package}
\alias{DAPS}
\alias{DAPS-package}
\title{DAPS: Deterministic Approximation of Posteriors for Bayesian fine-mapping}
\description{
DAPS implements the DAP-S algorithm for Bayesian genetic fine-mapping,
integrating SuSiE's variational approximation with deterministic posterior
computation to identify causal variants and construct interpretable
signal clusters.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://wnbo9.github.io/DAPS/}
  \item \url{https://github.com/wnbo9/DAPS}
  \item Report bugs at \url{https://github.com/wnbo9/DAPS/issues}
}

}
\author{
\strong{Maintainer}: Bo Wang \email{wnbo@umich.edu}

Authors:
\itemize{
  \item Xiaoquan Wen
}

}
\keyword{internal}
