% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daps.R
\name{daps}
\alias{daps}
\title{DAP-S fine-mapping with individual-level data}
\usage{
daps(
  X,
  y,
  L = min(10, ncol(X)),
  standardize = FALSE,
  prior = NULL,
  null_wt = NULL,
  proposal_thresh = 1e-06,
  grid = c(0.04, 0.16, 0.64),
  unmappable_effects = c("none", "inf", "ash"),
  twas_weight = FALSE,
  coverage = NULL,
  min_abs_corr = 0.5,
  dedup = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{X}{Genotype matrix (n x p) where n is the number of individuals and
p is the number of SNPs.}

\item{y}{Phenotype vector of length n.}

\item{L}{Maximum number of causal SNPs allowed in the model.  Default is
min(10, p).}

\item{standardize}{Standardize each column of X. Default is FALSE.}

\item{prior}{SNP-level prior weights, given as a vector of
length p. Default is NULL which corresponds to uninformative weights.}

\item{null_wt}{Null weight used in SuSiE, defined as the prior
probability of no causal variants in the region. Default is calculated
as \code{prod(1 - prior)}.}

\item{proposal_thresh}{Threshold for proposal acceptance. Default value
is 1e-6.}

\item{grid}{Grid of values for scaled effect size prior variance.
Default is c(0.04, 0.16, 0.64).}

\item{unmappable_effects}{Method for handling unmapped effects.
Options are "none", "inf", and "ash". Default is "none".}

\item{twas_weight}{Compute TWAS weights. Default is FALSE.}

\item{coverage}{Coverage leveel for credible set (e.g. 0.95). Default is
NULL, in which the algorithm returns signal clusters that are not
constrained by a fixed coverage level.}

\item{min_abs_corr}{Minimum absolute correlation allowed in a signal
cluster or credible set. Ddefault is 0.5, which corresponds to a
squared correlation of 0.25.}

\item{dedup}{Whether to remove duplicate effects
(effects with the same top 5 SNPs). Default is TRUE.}

\item{verbose}{Whether to print progress messages. Default is FALSE.}

\item{...}{Additional arguments passed to \code{\link{susie}}.}
}
\value{
A list of DAP-S fine-mapping results.
\describe{
\item{effect_pip}{A p*L matrix of SNP posterior probabilities in each signal}
\item{pip}{A p-vector of marginal SNP-level posterior inclusion probabilities}
\item{prior}{A p-vector of SNP prior probability of being causal}
\item{models}{A dataframe of high-probability models identified by DAP-S, together with their posterior probabilities}
\item{twas_weights}{A p-vector of TWAS weights for each SNP}
\item{info}{A list of information including signal clusters and stored model details}
\item{sets}{A list of signal clusters or credible sets identified by DAP-S with the specified coverage level}
\item{variants}{A variant-level dataframe for FastENLOC-style input}
}
}
\description{
DAP-S fine-mapping with individual-level genotype and phenotype data as
input.
}
\examples{
set.seed(1234)
n <- 1000
p <- 1000
beta <- rep(0, p)
beta[c(1, 200, 500, 800)] <- 1
X <- matrix(rnorm(n*p), nrow = n, ncol = p)
X <- scale(X, center = TRUE, scale = TRUE)
y <- X \%*\% beta + rnorm(n)
rst <- daps(X, y, L = 5)
rst$sets
daps_plot(rst)
}
