% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daps_plot.R
\name{daps_plot}
\alias{daps_plot}
\title{DAP-S PIP plot with cluster / credible-set annotation}
\usage{
daps_plot(res, band = 0)
}
\arguments{
\item{res}{A list of DAP-S fine-mapping results.}

\item{band}{Fraction of the row gap allowed for lifting between 0 and 1;
default 0.}
}
\value{
A ggplot object (patchwork layout).
}
\description{
DAP-S PIP plot with cluster / credible-set annotation
}
