% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daps.R
\name{daps_rss}
\alias{daps_rss}
\title{DAP-S fine-mapping with regression summary statistics}
\usage{
daps_rss(
  z = NULL,
  R,
  n = NULL,
  bhat = NULL,
  shat = NULL,
  var_y = NULL,
  estimate_residual_variance = FALSE,
  ...
)
}
\arguments{
\item{z}{A p-vector of single-SNP z-scores.}

\item{R}{A p x p LD matrix among the SNPs.}

\item{n}{Sample size.}

\item{bhat}{A p-vector of single-SNP effect size estimates.}

\item{shat}{A p-vector of standard errors for bhat. bhat and shat
together can be replaced by z.}

\item{var_y}{Phenotypic variance, defined as y'y/(n-1). When not
provided, the algorithm works on standardized data.}

\item{estimate_residual_variance}{Estimate residual variance in SuSiE.
Default is FALSE. If the in sample LD matrix is used, we recommend
setting this to TRUE.}

\item{...}{Other parameters to be passed to \code{\link{daps_suff_stat}}.}
}
\value{
A list of DAP-S fine-mapping results.
}
\description{
DAP-S fine-mapping using regression summary statistics to recover
sufficient statistics.
}
\examples{
z <- c(6,7)
R <- matrix(1,2,2)
rst <- daps_rss(z=z, R=R)
daps_plot(rst)
}
