% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daps_utils.R
\name{get_set}
\alias{get_set}
\title{DAP-S fine-mapping getting signal clusters or credible sets}
\usage{
get_set(output, coverage)
}
\arguments{
\item{output}{A list of DAP-S fine-mapping results.}

\item{coverage}{Desired coverage for credible sets.
If coverage is NULL, return signal clusters instead of credible sets.}
}
\value{
A list of signal clusters or credible sets.
}
\description{
DAP-S fine-mapping getting signal clusters or credible sets
}
