% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{lag}
\alias{case_when}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{case_when}}, \code{\link[dplyr]{lag}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}
}}

