% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_train.R
\name{train}
\alias{train}
\title{Train existing individual variable models.}
\usage{
train(daps, static, longitudinal)
}
\arguments{
\item{daps}{A \code{\link{daps}}-class object to add trained individual model fits to.}

\item{static}{A data frame with a unique numeric column \code{id}.}

\item{longitudinal}{A data frame with a numeric \code{id} column and an
\link[rlang:is_integerish]{integerish} \code{t} column.}
}
\value{
The input \code{daps} with the added component \code{$trained_fits}, an object
of class \code{daps_trained_fits}.
}
\description{
Runs the model calls that were added during individual model specification
(i.e., in \code{\link[=add_models]{add_models()}}).
}
\details{
The arguments \code{static} and \code{longitudinal} are validated and
\code{\link[dplyr:full_join]{dplyr::full_join()}}ed on the \code{id} column. This data set is inserted as the
\code{data} argument in each of the individual model calls, and each one is run.
Thus, these data should be just like the data that will ultimately be used
for simulation (i.e., in \code{\link[=simulate]{simulate()}}).
}
