% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daps_object.R
\name{daps}
\alias{daps}
\title{Create a new \code{daps} object.}
\usage{
daps()
}
\value{
An object of class \code{daps}.
}
\description{
The first step in the dynamically adaptive probabilistic system pipeline.
}
\details{
This should be piped into either \code{\link[=add_models]{add_models()}} or \code{\link[=add_metastate_model]{add_metastate_model()}}.
}
\section{Components}{

\code{daps}-class objects can contain the following components:
\itemize{
\item \strong{\code{daps_metastate_model}} added via \code{\link[=add_metastate_model]{add_metastate_model()}}. This contains the
state-transition framework.
\item \strong{\code{daps_model_table}} added via \code{\link[=add_models]{add_models()}}. This contains individual
variable models.
}
}

\examples{
daps()

}
