% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{daps_main}
\alias{daps_main}
\title{DAPS main function in Rcpp}
\usage{
daps_main(
  X,
  y,
  n,
  prior,
  proposal,
  proposal_thresh,
  grid,
  twas_weight,
  min_abs_corr,
  sigma2,
  sigma2_alpha
)
}
\arguments{
\item{X}{Genotype matrix (n x p)}

\item{y}{Phenotype vector (n x 1)}

\item{n}{Sample size (scalar)}

\item{prior}{Prior inclusion probabilities (p x 1)}

\item{proposal}{Proposal matrix (p x L) from SuSiE}

\item{proposal_thresh}{Proposal threshold (scalar)}

\item{grid}{Grid of scaled prior variance values (k x 1)}

\item{twas_weight}{Logical indicating whether to use TWAS weights}

\item{min_abs_corr}{Minimum absolute correlation for TWAS weights (scalar)}

\item{sigma2}{Residual variance estimate from SuSiE (scalar; used for inf or ash)}

\item{sigma2_alpha}{Estimated variance of small effect sizes (scalar; used for inf or ash)}
}
\value{
A list containing models and input data
}
\description{
DAPS main function in Rcpp
}
