% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daps.R
\name{daps_rss}
\alias{daps_rss}
\title{DAP-S fine-mapping with regression summary statistics}
\usage{
daps_rss(
  z = NULL,
  R,
  n = NULL,
  bhat = NULL,
  shat = NULL,
  var_y = NULL,
  prior = NULL,
  null_wt = NULL,
  estimate_residual_variance = FALSE,
  proposal_thresh = 1e-06,
  grid = c(0.04, 0.16, 0.64),
  unmappable_effects = c("none", "inf"),
  twas_weight = FALSE,
  coverage = NULL,
  min_abs_corr = 0.5,
  dedup = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{z}{A p-vector of single-SNP z-scores.}

\item{R}{A p x p LD matrix among the SNPs.}

\item{n}{Sample size.}

\item{bhat}{A p-vector of single-SNP effect size estimates.}

\item{shat}{A p-vector of standard errors for bhat. bhat and shat
together can be replaced by z.}

\item{var_y}{Phenotypic variance, defined as y'y/(n-1). When not
provided, the algorithm works on standardized data.}

\item{prior}{SNP-level prior weights, given as a vector of
length p. Default is NULL which corresponds to uninformative weights.}

\item{null_wt}{Null weight used in SuSiE, defined as the prior
probability of no causal variants in the region. Default is calculated
as \code{prod(1 - prior)}.}

\item{estimate_residual_variance}{Estimate residual variance in SuSiE.
Default is FALSE. If the in sample LD matrix is used, we recommend
setting this to TRUE.}

\item{proposal_thresh}{Threshold for proposal acceptance. Default value
is 1e-6.}

\item{grid}{Grid of values for scaled effect size prior variance.
Default is c(0.04, 0.16, 0.64).}

\item{unmappable_effects}{Method for handling unmapped effects.
Options are "none", "inf". Default is "none".}

\item{twas_weight}{Compute TWAS weights. Default is FALSE.}

\item{coverage}{Coverage level for credible set (e.g. 0.95). Default is
NULL, in which the algorithm returns signal clusters that are not
constrained by a fixed coverage level.}

\item{min_abs_corr}{Minimum absolute correlation allowed in a signal
cluster or credible set. Default is 0.5, which corresponds to a
squared correlation of 0.25.}

\item{dedup}{Whether to remove duplicate effects
(effects with the same top 5 SNPs). Default is TRUE.}

\item{verbose}{Whether to print progress messages. Default is FALSE.}

\item{...}{Other parameters to be passed to \code{\link{daps_suff_stat}}.}
}
\value{
A list of DAP-S fine-mapping results.
}
\description{
DAP-S fine-mapping using regression summary statistics to recover
sufficient statistics.
}
\examples{
z <- c(6,7)
R <- matrix(1,2,2)
rst <- daps_rss(z=z, R=R)
rst$pip
}
