% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{daps_ss_main}
\alias{daps_ss_main}
\title{DAPS_SS main function in Rcpp}
\usage{
daps_ss_main(
  XtX,
  Xty,
  yty,
  n,
  prior,
  proposal,
  proposal_thresh,
  grid,
  twas_weight,
  min_abs_corr,
  sigma2,
  sigma2_alpha
)
}
\arguments{
\item{XtX}{Cross-product of genotype matrix (p x p)}

\item{Xty}{Cross-product of genotype and phenotype vector (p x 1)}

\item{yty}{Cross-product of phenotype vector (scalar)}

\item{n}{Sample size (scalar)}

\item{prior}{Prior inclusion probabilities (p x 1)}

\item{proposal}{Proposal matrix (p x L) from SuSiE}

\item{proposal_thresh}{Proposal threshold (scalar)}

\item{grid}{Grid of scaled prior variance values (k x 1)}

\item{twas_weight}{Logical indicating whether to use TWAS weights}

\item{min_abs_corr}{Minimum absolute correlation for TWAS weights (scalar)}

\item{sigma2}{Residual variance estimate from SuSiE (scalar; used for inf or ash)}

\item{sigma2_alpha}{Estimated variance of small effect sizes (scalar; used for inf or ash)}
}
\value{
A list containing models and input data
}
\description{
DAPS_SS main function in Rcpp
}
